Add golangci-lint to workflow
By introducing a linter into your project, you can reduce the cost of code checking.
The golangci-lint we will introduce this time is a tool that runs linters at high speed and is well known in the Go community.

The default linter is here: https://golangci-lint.run/usage/linters/#enabled-by-default